# Use rb_utf8_str_new/rb_utf8_str_new_cstr to create UTF8 string