--- a/ext/oj/string_writer.c
+++ b/ext/oj/string_writer.c
@@ -469,9 +469,7 @@ static VALUE str_writer_reset(VALUE self) {
 static VALUE str_writer_to_s(VALUE self) {
     StrWriter sw;
     TypedData_Get_Struct(self, struct _strWriter, &oj_string_writer_type, sw);
-    VALUE rstr = rb_str_new(sw->out.buf, sw->out.cur - sw->out.buf);
-
-    return oj_encode(rstr);
+    return rb_utf8_str_new(sw->out.buf, sw->out.cur - sw->out.buf);
 }
 
 /* Document-method: as_json