@Watson1978 2025-01-04
Hmm, It's still lagging behind json gem...

```ruby
require 'bundler/inline'
gemfile do
  source 'https://rubygems.org'
  gem 'benchmark-ips'
  gem 'oj'
  gem 'json'
end

# https://github.com/miloyip/nativejson-benchmark/blob/master/data/twitter.json
json = File.read('twitter.json')

Benchmark.ips do |x|
  x.time = 10
  x.report('JSON.parse') { JSON.parse(json) }
  x.report('Oj.load compat') { Oj.load(json, mode: :compat) }
  x.compare!
end
```

```
Warming up --------------------------------------
          JSON.parse    70.000 i/100ms
      Oj.load compat    65.000 i/100ms
Calculating -------------------------------------
          JSON.parse    739.810 (± 3.1%) i/s    (1.35 ms/i) -      7.420k in  10.040792s
      Oj.load compat    655.715 (± 2.9%) i/s    (1.53 ms/i) -      6.565k in  10.021079s

Comparison:
          JSON.parse:      739.8 i/s
      Oj.load compat:      655.7 i/s - 1.13x  slower
```