@ohler55 2025-01-05
The biggest performance bottleneck is the call to the objects `to_json` or `as_json` methods. I'm not sure what can be done to avoid those calls.